DoctrineEventStore implements GapDetection itself
