+++ src/Common/EventStore/Integration/Symfony/CleanUpEventStoreCommand.php
@@ -14,7 +14,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 /**
- * If there are multiple streaming process, this command can be used to safely
+ * If there are multiple streaming processes, this command can be used to safely
  * clean up the EventStore based on the given EventStorePointers.
  */
 final class CleanUpEventStoreCommand extends Command
